providers: rename vagrant-virtualbox to virtualbox
This allows us to avoid using specific oem-ids for vagrant provisioned
machines and use more ids
